Implement UpdateReleases as Symfony command
