+++ Makefile
@@ -14,7 +14,7 @@ init: start
 update-data:
 	${APP-RUN} bin/console app:update:mirrors
 	${APP-RUN} bin/console app:update:news
-	${APP-RUN} bin/console app:update releases
+	${APP-RUN} bin/console app:update:releases
 	${APP-RUN} bin/console app:update packages
 	${APP-RUN} bin/console app:update:pkgstats
